requirements.txt
fastapi~=0.83.0
psutil~=5.9.4
win32gui~=221.6
numpy~=1.19.5
pillow~=8.4.0
